pyproject.toml
[project]
name = "registry-manager"
version = "0.1.0"
description = "Lightweight private Docker registry management platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.121.1",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[tool.uv]
package = false
